MOD: 3계층 레이어 적용 및 MVP 패턴으로 리팩터링
