--- Main.java
@@ -1,9 +1,0 @@
-import screen.MainScreen;
-import theme.ThemeColor;
-
-public class Main {
-    public static void main(String[] args) {
-        ThemeColor.applyThemeDark("theme/dark-ui.properties");
-        MainScreen mainScreen = new MainScreen();
-    }
-}
