--- manager/Manageable.java
@@ -1,11 +1,0 @@
-package manager;
-
-import java.util.Scanner;
-
-public interface Manageable {
-
-    void read(Scanner scanner);
-    String toString();
-    boolean matches(String kwd);
-    String getId();
-}
