--- screen/RecipeDialog.java
@@ -1,4 +1,0 @@
-package screen;
-
-public class RecipeDialog {
-}
