+++ src/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package src.repository;
+
+import src.entity.User;
+
+import java.io.IOException;
+import java.util.List;
+
+public interface UserRepository {
+    List<User> loadAll();
+    void save(User user) throws IOException;
+}
