+++ src/screen/navigation/NavigationHandler.java
@@ -1,0 +1,7 @@
+package src.screen.navigation;
+
+public interface NavigationHandler {
+    void displayHomeScreen();
+    void displayCategoryScreen();
+    void displayPlannerScreen();
+}
